refactor: thread 폴더추가
