+++ src/pages/Detail/components/thread/UpdateDeleteDialog.tsx
